+++ openhands/app_server/app_conversation/app_conversation_router.py
@@ -69,7 +69,7 @@ from openhands.app_server.utils.dependencies import get_dependencies
 from openhands.app_server.utils.docker_utils import (
     replace_localhost_hostname_for_docker,
 )
-from openhands.sdk.context.skills import KeywordTrigger, TaskTrigger
+from openhands.sdk.skills import KeywordTrigger, TaskTrigger
 from openhands.sdk.workspace.remote.async_remote_workspace import AsyncRemoteWorkspace
 
 # Handle anext compatibility for Python < 3.10
